+++ CHANGES.md
@@ -1,12 +1,19 @@
 # Change Log
 
-## v0.9.6
+## v0.11.0
 
 Added new actions:
 * `vsphere.vm_hw_hdds_get` - Retrieve a list of HDD information from the given VM
 * `vsphere.vm_hw_scsi_controllers_get` - Retrieve a list of SCSI controllers on the given VM
 
 Contributed by John Schoewe (Encore Technologies).
+
+## v0.10.0
+
+Added new action:
+* `vsphere.vm_snapshots_get` - Returns detailed information about the snapshots.
+
+Contributed by Lionel Seydoux (SPIE ICS).
 
 ## v0.9.5
 
@@ -271,4 +278,3 @@ Clean up of duplicate actions. Addition of moid retrieval function.
 ## V0.2
 
 Addition of vm_hw atomic actions to create basic Virtual Machine.
-